rollback YouTube functionality back to Web API to handle livestream announcements/community posts etc.
